+++ lib/masking/errors.rb
@@ -5,5 +5,6 @@ module Masking
     class ConfigFileDoesNotExist < Error; end
     class ConfigFileIsNotFile < Error; end
     class ConfigFileIsNotValidYaml < Error; end
+    class InsertStatementParseError < Error; end
   end
 end