+++ spec/fixtures/insert_statement/without_complete_insert_option.sql
@@ -1,0 +1,1 @@
+INSERT INTO `dummy` VALUES (1,'Super Chikahiro','kibitan@example.com','password_digest','2018-03-14 00:00:00','2018-03-29 00:00:00'),(2,'Super Tokoro','kibitan++@example.com','password_digest2','2018-04-01 00:00:00','2018-04-03 12:00:00');