+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `masking --version` option [#31](https://github.com/kibitan/masking/pull/31)
+- add friendly parse error message [#26](https://github.com/kibitan/masking/pull/26)
 
 ## [v0.0.3] - 2019-07-07